Add and test extracting 16-bit function from frame
It appears that all frames have a two-byte function code after the
length, which is the same for both request and reply frames, so extract
this as part of the frame header.

Add a test for the returned header value and tidy up the rest of the
tests to not require a helper method.