Implement creating new frames as well as reading
Add a constructor and an implicit conversion from Frame to vector of
bytes so that Frame objects can be passed cleanly to the WriteBytes
method of the Serial class.  Also add a test for the conversion and
update the boiler-logger executable to use this to send identify and
status requests.