Refactor in 'factory' Message class above Identify
To make handling the different messages passed back more easily, extract
some of the methods in the IdentifyMessage class into a Message parent
class which contains a 'FromFrame' function to choose the relevant
sub-class.

Also change the frame function type to an enumeration to avoid
sprinkling magic numbers all through the code.