Add a function for reading from boiler EEPROM
Introduce a function to the boiler class to read a block of EEPROM
memory from the boiler.  As the function code to read a block of EEPROM
uses the second byte as the block number, update the Frame class to
handle switching the values around the correct way and back when
decooding and encoding frames.