Reduce logging frequency of ever changing values
To reduce logging volume of temperatures, ionisation, etc, implement a
basic per-attribute frequency limiting mechanism.  Each changed
attribute value is only written to the log if it averages less than one
log message every 30 seconds of the duration of the application.  This
allows changes to infrequently changed values like state to still all be
logged.

Also increase the sampling frequency now that the limiting is in place.