Set serial options to prevent data corruption
By default, some options can be enabled which modify the input and
output data over the serial port.  Disable these to prevent the data
becoming corrupted.

Also add some more information to the frame length and incorrect
checksum exceptions in the Frame class which were useful to debug this
issue.