Only apply log rate limiting to numeric values
Values which contain numbers (which are the values that tend to change
constantly) can be recognised by checking for '[' in the value name
(units are always shown in square brackets).  Use that information to
only apply log rate limiting to numeric values.

Also double the maximum log rate for numeric values to one a minute
instead of one every 30 seconds.