Move author copyright to individual files
Instead of trying to handle the list of authors separately to the file
copyright, remove AUTHORS.md and update the copyright in the files to
have the correct names and years.  Any future contributors can add their
own copyright notices directly to the files they contribute to.

Also take the opportunity to reduce the size of the headers by removing
the GPL boilerplate from each file and replacing it with a
SPDX-License-Identifier line.