Fix incorrectly sized memcpy when building a Frame
When converting a Frame object to a raw byte vector, the code currently
copies sizeof(header) bytes from the trailer to the end of the vector
instead of sizeof(trailer).  As header is 6 bytes long but trailer is
only 2 this means that four bytes of stack memory after trailer are
being copied to the stack memory after the raw vector.  Happily this
doesn't appear to have been having any ill effects up to now, presumably
just due to luck (and padding between the stack elements?).

This was flagged up as a warning in GCC 12.2.  Correct the variable
passed to sizeof() so that the correct amount of memory is copied
instead.