Also handle inline text found in config files
Some text found in the xml configuration files isn't a reference to a
language file string (attribute ending in .nr), occasionally it is
defined directly in the config file (in an attribute with the same
prefix but the suffix .txt).

Extend convert.pl to fall back to a .txt attribute content if a .nr
attribute does not exist on the XML node.